feat: GridCollectionViewCell 구현

+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -11,9 +11,9 @@ struct Product: Decodable, Identifiable, Equatable {
     let productDescription: String
     let thumbnailURL: String
     let currency: Currency
-    let price: Int
-    let bargainPrice: Int
-    let discountedPrice: Int
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
     let stock: Int
     let createdAt: String
     let issuedAt: String
